requirements.txt
customtkinter
packaging
Pillow
openai==0.28
opencv-python
ffmpeg
pyaudio
moviepy